% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlmoderator-package.R
\docType{package}
\name{mlmoderator-package}
\alias{mlmoderator}
\alias{mlmoderator-package}
\title{mlmoderator: Probing and Visualizing Multilevel Interaction Effects}
\description{
\code{mlmoderator} provides a unified workflow for estimating, probing, and
visualizing multilevel moderation effects from mixed-effects models fitted
with \code{lme4::lmer()}. It supports simple slopes analysis, Johnson---Neyman
intervals, publication-ready interaction plots, and grand- or group-mean
centering.
\subsection{Core functions}{\tabular{ll}{
   Function \tab Description \cr
   \code{mlm_center()} \tab Grand- or group-mean center variables \cr
   \code{mlm_probe()} \tab Compute simple slopes at selected moderator values \cr
   \code{mlm_jn()} \tab Johnson-Neyman significance regions \cr
   \code{mlm_plot()} \tab Publication-ready interaction plot \cr
   \code{mlm_summary()} \tab Consolidated moderation report \cr
   \code{mlm_variance_decomp()} \tab Decompose slope uncertainty into fixed + random components \cr
   \code{mlm_surface()} \tab Slope surface heatmap over the full predictor x moderator space \cr
   \code{mlm_sensitivity()} \tab Robustness of interaction to ICC shift and omitted confounders \cr
}

}

\subsection{Typical workflow}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(mlmoderator)
library(lme4)

data(school_data)

# 1. Center variables
dat <- mlm_center(school_data, vars = "ses", cluster = "school", type = "group")

# 2. Fit model
mod <- lmer(math ~ ses * climate + gender + (1 + ses | school), data = dat)

# 3. Probe interaction
mlm_probe(mod, pred = "ses", modx = "climate")

# 4. Johnson-Neyman interval
mlm_jn(mod, pred = "ses", modx = "climate")

# 5. Plot
mlm_plot(mod, pred = "ses", modx = "climate")

# 6. Summary report
mlm_summary(mod, pred = "ses", modx = "climate")
}\if{html}{\out{</div>}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/causalfragility-lab/mlmoderator}
  \item Report bugs at \url{https://github.com/causalfragility-lab/mlmoderator/issues}
}

}
\author{
\strong{Maintainer}: Subir Hait \email{haitsubi@msu.edu} (\href{https://orcid.org/0009-0004-9871-9677}{ORCID})

}
\keyword{internal}
